Extract firmware-job follow endpoints into follow.py
Sixth structural-split PR for controllers/firmware/controller.py
(follows #733 persistence + #735 cli + #739 lifecycle + #740
factories + #745 runner). Pulls the WS streaming machinery —
follow_job and follow_jobs — into a new follow.py submodule.

The two methods are heavily used by tests (test_follow_job_race,
test_follow_jobs_race, test_branches_coverage) but only via the
public @api_command-decorated entry points; the closures inside
(_send_initial / _handle_event) are not patched. The decorator
+ kwargs signature stays on the controller method so the WS
dispatch table keeps working unchanged; only the bodies move.

Changes:

* New follow.py (228 lines) — two free functions, one per
  follower kind. Coherent unit: same race-against-stream_events
  reasoning, same StreamControls shape, same bus event types
  (JOB_OUTPUT + TERMINAL_JOB_EVENTS for single-job, full
  lifecycle set for all-jobs).
* controller.py — 1230 -> 1037 lines (-193). Bodies replaced
  with thin delegates; ruff auto-removed 5 now-unused imports
  (StreamControls, stream_events, StreamEvent, TERMINAL_JOB_EVENTS,
  the inline Event type-only import).

Combined arc (#733 + #735 + #739 + #740 + #745 + this):
controller.py 2083 -> 1037 lines (-50%, halved).

All firmware tests pass; ruff + mypy clean. No production-side
behaviour change.
